requirements.txt
flake8
line_profiler
coverage
memory_profiler
pillow